third_party/blink/web_tests/virtual/async-script-scheduling-finished-parsing/wpt_internal/async-script-scheduling/execution-order-expected.txt

This is a testharness.js-based test.
FAIL Async Script Execution Order Uncaught Error: assert_equals: expected "Async;Async;Async;inline1;external;inline2;DOMContentLoaded" but got "Async;inline1;external;inline2;DOMContentLoaded;Async;Async"
Harness: the test ran to completion.
